replace print with logger
